Rename addon.js register.js
So that it's clearer to folks that they shouldn't put this in the addons
file:

import 'happo-plugin-storybook/register';
